Remove callable firewall listeners support
