+++ UPGRADE-8.0.md
@@ -326,6 +326,10 @@ Security
   +}
   ```
 
+ * Remove callable firewall listeners support, extend `AbstractListener` or implement `FirewallListenerInterface` instead
+ * Remove `AbstractListener::__invoke`
+ * Remove `LazyFirewallContext::__invoke()`
+
 Serializer
 ----------
