+++ src/Symfony/Bundle/SecurityBundle/Debug/WrappedLazyListener.php
@@ -12,10 +12,13 @@
 namespace Symfony\Bundle\SecurityBundle\Debug;
 
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
 use Symfony\Component\Security\Core\Exception\LazyResponseException;
+use Symfony\Component\Security\Http\Authenticator\Debug\TraceableAuthenticatorManagerListener;
 use Symfony\Component\Security\Http\Firewall\AbstractListener;
 use Symfony\Component\Security\Http\Firewall\FirewallListenerInterface;
+use Symfony\Component\VarDumper\Caster\ClassStub;
 
 /**
  * Wraps a lazy security listener.
@@ -26,7 +29,10 @@ use Symfony\Component\Security\Http\Firewall\FirewallListenerInterface;
  */
 final class WrappedLazyListener extends AbstractListener
 {
-    use TraceableListenerTrait;
+    private ?Response $response = null;
+    private FirewallListenerInterface $listener;
+    private ?float $time = null;
+    private ClassStub $stub;
 
     public function __construct(FirewallListenerInterface $listener)
     {
@@ -53,5 +59,22 @@ final class WrappedLazyListener extends AbstractListener
         }
 
         $this->response = $event->getResponse();
+    }
+
+    public function getInfo(): array
+    {
+        return [
+            'response' => $this->response,
+            'time' => $this->time,
+            'stub' => $this->stub ??= new ClassStub($this->listener instanceof TraceableAuthenticatorManagerListener ? $this->listener->getAuthenticatorManagerListener()::class : $this->listener::class),
+        ];
+    }
+
+    /**
+     * Proxies all method calls to the original listener.
+     */
+    public function __call(string $method, array $arguments): mixed
+    {
+        return $this->listener->{$method}(...$arguments);
     }
 }
