+++ src/Symfony/Bundle/SecurityBundle/Security/FirewallContext.php
@@ -24,7 +24,7 @@ use Symfony\Component\Security\Http\Firewall\LogoutListener;
 class FirewallContext
 {
     /**
-     * @param iterable<mixed, callable> $listeners
+     * @param iterable<mixed, FirewallListenerInterface> $listeners
      */
     public function __construct(
         private iterable $listeners,
@@ -40,7 +40,7 @@ class FirewallContext
     }
 
     /**
-     * @return iterable<mixed, FirewallListenerInterface|callable>
+     * @return iterable<mixed, FirewallListenerInterface>
      */
     public function getListeners(): iterable
     {
