+++ src/Symfony/Bundle/SecurityBundle/composer.json
@@ -22,7 +22,6 @@
         "symfony/clock": "^7.4|^8.0",
         "symfony/config": "^7.4|^8.0",
         "symfony/dependency-injection": "^7.4|^8.0",
-        "symfony/deprecation-contracts": "^2.5|^3",
         "symfony/event-dispatcher": "^7.4|^8.0",
         "symfony/http-kernel": "^7.4|^8.0",
         "symfony/http-foundation": "^7.4|^8.0",
