+++ src/Symfony/Component/Security/Http/Firewall/AbstractListener.php
@@ -20,18 +20,6 @@ use Symfony\Component\HttpKernel\Event\RequestEvent;
  */
 abstract class AbstractListener implements FirewallListenerInterface
 {
-    /**
-     * @deprecated since Symfony 7.4, to be removed in 8.0
-     */
-    final public function __invoke(RequestEvent $event): void
-    {
-        trigger_deprecation('symfony/security-http', '7.4', 'The "%s()" method is deprecated since Symfony 7.4 and will be removed in 8.0.', __METHOD__);
-
-        if (false !== $this->supports($event->getRequest())) {
-            $this->authenticate($event);
-        }
-    }
-
     public static function getPriority(): int
     {
         return 0; // Default
