+++ src/Symfony/Component/Security/Http/FirewallMapInterface.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Security\Http;
 
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Security\Http\Firewall\ExceptionListener;
+use Symfony\Component\Security\Http\Firewall\FirewallListenerInterface;
 use Symfony\Component\Security\Http\Firewall\LogoutListener;
 
 /**
@@ -35,7 +36,7 @@ interface FirewallMapInterface
      * If there is no logout listener, the third element of the outer array
      * must be null.
      *
-     * @return array{iterable<mixed, callable>, ExceptionListener, LogoutListener}
+     * @return array{iterable<mixed, FirewallListenerInterface>, ExceptionListener, LogoutListener}
      */
     public function getListeners(Request $request): array;
 }
